Refactor story definitions and rename table patterns
