--- .storybook/stories/DataFiltering/TableFilteringLargeExample.tsx
@@ -1,17 +1,0 @@
-import { TableFilteringLarge } from './TableFilteringLarge';
-import { useSortAndFilter } from './lib/useSortAndFilter';
-import { useData } from './lib/utils';
-import { DataProvider, SortAndFilterProvider } from './lib/contexts';
-
-export const TableFilteringLargeExample = () => {
-	const sortAndFilter = useSortAndFilter();
-	const { filters, pagination, sort } = sortAndFilter;
-	const data = useData({ filters, pagination, sort });
-	return (
-		<SortAndFilterProvider value={sortAndFilter}>
-			<DataProvider value={data}>
-				<TableFilteringLarge />
-			</DataProvider>
-		</SortAndFilterProvider>
-	);
-};
