+++ docs/content/patterns/filtering-a-dataset/index.mdx
@@ -3,7 +3,7 @@ title: Filtering a dataset
 description: Filters allows a user to refine the data they want to display in a large dataset by removing any items that do not match their criteria.
 figmaTemplateNodeId: 18586-56265
 githubTemplatePath: /.storybook/stories/DataFiltering
-storybookPath: /story/patterns-data-filtering-and-sorting--small
+storybookPath: /story/patterns-tables--filter-small
 relatedComponents:
   ['filter-drawer', 'filter-sidebar', 'table', 'card', 'pagination']
 ---
@@ -104,7 +104,7 @@ Small is intended for 1-2 filters that are visible and sit above the dataset.
 
 The small pattern is typically comprised of primary filters.
 
-[View Storybook preview](/storybook/index.html?path=/story/patterns-data-filtering-and-sorting--small)
+[View Storybook preview](/storybook/index.html?path=/story/patterns-tables--filter-small)
 
 <ImageWithBorder
 	src="/img/patterns/filtering-a-dataset-small.png"
@@ -119,7 +119,7 @@ Medium is intended to display 3-6 filters in an accordion that is opened and clo
 
 Tags are added below the filters to show which are active and give the user an easy way to clear them.
 
-[View Storybook preview](/storybook/index.html?path=/story/patterns-data-filtering-and-sorting--medium)
+[View Storybook preview](/storybook/index.html?path=/story/patterns-tables--filter-medium)
 
 <ImageWithBorder
 	src="/img/patterns/filtering-a-dataset-medium.png"
@@ -150,7 +150,7 @@ The [Drawer](/components/drawer) component should contain a total of 4 actions:
 	alt="Screenshot of the large table filtering pattern"
 />
 
-[View Storybook preview](/storybook/index.html?path=/story/patterns-data-filtering-and-sorting--large)
+[View Storybook preview](/storybook/index.html?path=/story/patterns-tables--filter-large)
 
 ## Filtering a list of cards
 
@@ -158,4 +158,4 @@ This pattern is used to display a collection of filterable links ([Cards](/compo
 
 We use a two-column layout featuring the [Filter sidebar](/components/filter-sidebar) component, which contains all of the filter options. It collapses on mobile devices, and displays a number of active filter options in the header.
 
-[View Storybook preview](/storybook/index.html?path=/story/patterns-data-filtering-and-sorting--list)
+[View Storybook preview](/storybook/index.html?path=/story/http://localhost:6006/?path=/story/patterns-search-results--list)
