+++ packages/react/src/table/docs/overview.mdx
@@ -280,7 +280,7 @@ You can use custom widths based on the expected length of the data under each co
 
 You can use the `TableHeaderSortable` component to make columns sortable. This component is a wrapper around the `TableHeader` component and adds the necessary aria attributes to make the column sortable.
 
-You can see a working example of this component in the [Table Filtering Patterns](/storybook/index.html?path=/story/patterns-data-filtering-and-sorting--medium).
+You can see a working example of this component in the [Table Filtering Patterns](/storybook/index.html?path=/story/patterns-tables--filter-medium).
 
 ```jsx live
 <TableWrapper>
